+++ packages/lib-sourcify/src/index.ts
@@ -10,6 +10,7 @@ export type ILibSourcifyLogger = ILogger;
 export * from './Compilation/AbstractCompilation';
 export * from './Compilation/SolidityCompilation';
 export * from './Compilation/VyperCompilation';
+export * from './Compilation/YulCompilation';
 export * from './Compilation/PreRunCompilation';
 export * from './Compilation/CompilationTypes';