+++ services/server/src/server/apiv1/verification/private/stateless/private.stateless.handlers.ts
@@ -14,9 +14,8 @@ import type {
 import {
   createMetadataContractsFromFiles,
   Verification,
-  SolidityCompilation,
-  VyperCompilation,
   splitFullyQualifiedName,
+  CompilationError,
 } from "@ethereum-sourcify/lib-sourcify";
 import {
   BadRequestError,
@@ -33,6 +32,7 @@ import SourcifyChainMock from "../../../../services/utils/SourcifyChainMock";
 import { getCreatorTx } from "../../../../services/utils/contract-creation-util";
 import type { CustomReplaceMethod } from "./customReplaceMethods";
 import { REPLACE_METHODS } from "./customReplaceMethods";
+import { createCompilationFromJsonInput } from "../../../../services/utils/compilation";
 
 export async function verifyDeprecated(
   req: LegacyVerifyRequest,
@@ -236,25 +236,25 @@ export async function replaceContract(
           "jsonInput, compilerVersion and compilationTarget are required when forceCompilation is true",
         );
       }
-      if (jsonInput?.language === "Solidity") {
-        compilation = new SolidityCompilation(
-          solc,
+      try {
+        compilation = createCompilationFromJsonInput(
+          { solc, vyper },
           compilerVersion,
-          jsonInput as SolidityJsonInput,
+          jsonInput,
           compilationTarget,
         );
-      } else if (jsonInput?.language === "Vyper") {
-        compilation = new VyperCompilation(
-          vyper,
-          compilerVersion,
-          jsonInput as VyperJsonInput,
-          compilationTarget,
-        );
-      } else {
-        throw new BadRequestError(
-          "Invalid language. Only Solidity and Vyper are supported",
-        );
+      } catch (err: any) {
+        if (
+          err instanceof CompilationError &&
+          err.code === "invalid_language"
+        ) {
+          throw new BadRequestError(
+            "Invalid language. Only Solidity and Vyper are supported",
+          );
+        }
+        throw err;
       }
+
       await compilation.compile();
     }